Made new loading activity
