+++ .idea/misc.xml
@@ -28,6 +28,7 @@
         <entry key="..\:/Users/Julian Chan/OneDrive/Documents/SCHOOL/2022_SUMMER/CMPT_362/Project/app/src/main/res/xml/options_settings.xml" value="0.3328125" />
         <entry key="..\:/Users/Julian Chan/OneDrive/Documents/SCHOOL/2022_SUMMER/CMPT_362/Project/app/src/main/res/xml/preference.xml" value="0.5369791666666667" />
         <entry key="..\:/Users/Julian Chan/OneDrive/Documents/SCHOOL/2022_SUMMER/CMPT_362/Project/app/src/main/res/xml/preference_settings.xml" value="0.3328125" />
+        <entry key="..\:/Users/Shashank/Desktop/Summer 2022/CMPT362/Project/BlueRadar/app/src/main/res/layout/activity_loading.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/Shashank/Desktop/Summer 2022/CMPT362/Project/BlueRadar/app/src/main/res/layout/activity_main.xml" value="0.264" />
         <entry key="..\:/Users/Shashank/Desktop/Summer 2022/CMPT362/Project/BlueRadar/app/src/main/res/layout/activity_splash_screen.xml" value="0.4036458333333333" />
         <entry key="..\:/Users/Shashank/Desktop/Summer 2022/CMPT362/Project/BlueRadar/app/src/main/res/layout/app_bar_main.xml" value="0.40260416666666665" />
