+++ app/src/main/AndroidManifest.xml
@@ -3,7 +3,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     package="ca.sfu.BlueRadar"
     tools:ignore="CoarseFineLocation">
-
     <!-- Request legacy Bluetooth permissions on older devices. -->
     <uses-permission
         android:name="android.permission.BLUETOOTH"
@@ -51,6 +50,10 @@
         android:supportsRtl="true"
         android:theme="@style/MyAppTheme"
         tools:targetApi="31">
+        <activity
+            android:name=".LoadingActivity"
+            android:exported="false" />
+
         <meta-data
             android:name="com.google.android.actions"
             android:resource="@xml/file_paths" />
@@ -60,18 +63,16 @@
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
         <activity
             android:name=".about.AboutApplication"
             android:exported="false" />
-
         <activity
             android:name=".about.AboutDevelopers"
             android:exported="false" />
-
         <activity
             android:name=".navigation.NavigationActivity"
             android:exported="false"
@@ -85,7 +86,6 @@
              sign the APK for publishing.
              You can define the keys for the debug and release targets in src/debug/ and src/release/.
         -->
-
         <meta-data
             android:name="com.google.android.geo.API_KEY"
             android:value="@string/google_maps_key" />
@@ -116,8 +116,7 @@
 
         <activity
             android:name=".MainActivity"
-            android:exported="true">
-        </activity>
+            android:exported="true"></activity>
 
         <provider
             android:name="androidx.core.content.FileProvider"
